Added a userguide help button
removed user guide help button
added user help guide into the cascade drop down bar at the top
removed redundant code and cleaned up parts

26/04
Read the UID from a dropdown menu
Problem:
It selects only 1 value rather than the selected value, regardless of which is selected.
Attempted Tracing to allow it to work.
TO DO:
- Get code to split the dataset based on iPhoneUID

27/04
Individual dataset can be selected

28/04
Dataset is now splitting
UserID is implemented successfully!
Data is now needing to be put into an array that matches the UID so it can be plotted.

To Do
Make it work